+++ CHANGELOG.md
@@ -53,6 +53,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1195](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1195)
 - Correct Layering of Items on the Page when Dialogs are Utilized
   [GEOD-1189](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1189)
+- Detail Page Info Boxes
+  [GEOD-1197](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1197)
 
 ## [3.0.0-alpha.4] - 2018-10-01